# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Nifty Pay - Global Money Transfer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  avatarInitials: string;
  memberSince: string;
}

export type TransactionStatus = typeof TransactionStatus[keyof typeof TransactionStatus];


export const TransactionStatus = {
  pending: 'pending',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface Transaction {
  id: number;
  fromCurrency: string;
  toCurrency: string;
  fromAmount: number;
  toAmount: number;
  exchangeRate: number;
  fee: number;
  status: TransactionStatus;
  recipientName: string;
  recipientCountry: string;
  recipientFlag?: string;
  /** @nullable */
  note?: string | null;
  createdAt: string;
}

export interface DashboardSummary {
  totalBalanceUsd: number;
  totalWallets: number;
  completedTransfers: number;
  totalVolume: number;
  recentTransactions: Transaction[];
}

export interface Wallet {
  id: number;
  currencyCode: string;
  currencyName: string;
  balance: number;
  flag: string;
  isCrypto: boolean;
  /** @nullable */
  cryptoSymbol?: string | null;
}

export interface TopUpInput {
  amount: number;
}

export interface TransactionInput {
  fromWalletId: number;
  toCurrencyCode: string;
  fromAmount: number;
  recipientName: string;
  recipientCountry: string;
  note?: string;
}

export interface ExchangeRateResult {
  from: string;
  to: string;
  rate: number;
  fee: number;
  feeAmount: number;
  inverseRate: number;
}

export interface Country {
  code: string;
  name: string;
  currencyCode: string;
  currencyName: string;
  flag: string;
  transferFee: number;
  estimatedTime: string;
  popular?: boolean;
}

export interface CurrencyVolume {
  currencyCode: string;
  flag: string;
  totalVolume: number;
  count: number;
}

export interface TransactionStats {
  byCurrency: CurrencyVolume[];
  successRate: number;
  avgTransferTime: string;
}

export type GetTransactionsParams = {
status?: GetTransactionsStatus;
limit?: number;
};

export type GetTransactionsStatus = typeof GetTransactionsStatus[keyof typeof GetTransactionsStatus];


export const GetTransactionsStatus = {
  pending: 'pending',
  completed: 'completed',
  failed: 'failed',
} as const;

export type GetExchangeRatesParams = {
from: string;
to?: string;
};