# lib/api-zod/src/generated/types/country.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Nifty Pay - Global Money Transfer API
 * OpenAPI spec version: 0.1.0
 */

export interface Country {
  code: string;
  name: string;
  currencyCode: string;
  currencyName: string;
  flag: string;
  transferFee: number;
  estimatedTime: string;
  popular?: boolean;
}