# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Nifty Pay - Global Money Transfer API
 * OpenAPI spec version: 0.1.0
 */
import type { Transaction } from './transaction';

export interface DashboardSummary {
  totalBalanceUsd: number;
  totalWallets: number;
  completedTransfers: number;
  totalVolume: number;
  recentTransactions: Transaction[];
}