# lib/api-zod/src/generated/types/exchangeRateResult.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Nifty Pay - Global Money Transfer API
 * OpenAPI spec version: 0.1.0
 */

export interface ExchangeRateResult {
  from: string;
  to: string;
  rate: number;
  fee: number;
  feeAmount: number;
  inverseRate: number;
}