# lib/api-zod/src/generated/types/transactionStats.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Nifty Pay - Global Money Transfer API
 * OpenAPI spec version: 0.1.0
 */
import type { CurrencyVolume } from './currencyVolume';

export interface TransactionStats {
  byCurrency: CurrencyVolume[];
  successRate: number;
  avgTransferTime: string;
}