# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Nifty Pay - Global Money Transfer API
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  avatarInitials: string;
  memberSince: string;
}