# lib/api-zod/src/generated/types/wallet.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Nifty Pay - Global Money Transfer API
 * OpenAPI spec version: 0.1.0
 */

export interface Wallet {
  id: number;
  currencyCode: string;
  currencyName: string;
  balance: number;
  flag: string;
  isCrypto: boolean;
  /** @nullable */
  cryptoSymbol?: string | null;
}