# lib/db/src/schema/transactions.ts
import { pgTable, serial, text, numeric, timestamp } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const transactionsTable = pgTable("transactions", {
  id: serial("id").primaryKey(),
  fromCurrency: text("from_currency").notNull(),
  toCurrency: text("to_currency").notNull(),
  fromAmount: numeric("from_amount", { precision: 18, scale: 4 }).notNull(),
  toAmount: numeric("to_amount", { precision: 18, scale: 4 }).notNull(),
  exchangeRate: numeric("exchange_rate", { precision: 18, scale: 6 }).notNull(),
  fee: numeric("fee", { precision: 18, scale: 4 }).notNull().default("0"),
  status: text("status").notNull().default("pending"),
  recipientName: text("recipient_name").notNull(),
  recipientCountry: text("recipient_country").notNull(),
  recipientFlag: text("recipient_flag").notNull(),
  note: text("note"),
  createdAt: timestamp("created_at", { withTimezone: true }).notNull().defaultNow(),
});

export const insertTransactionSchema = createInsertSchema(transactionsTable).omit({ id: true, createdAt: true });
export type InsertTransaction = z.infer<typeof insertTransactionSchema>;
export type Transaction = typeof transactionsTable.$inferSelect;